Modification of setPlaybackStatus events
The seekTo function in AwesomePlayer can be called both when playing
content but also when the content is paused. In the old solution
the events to the DRM FW were the same for both these cases.
This causes too many DRM rights to be consumed in the case where the
song is paused and you seek to the beginning.

Change-Id: I19d8926745595d55962f27ae2bad2c00d42a15eb